Extract room request handling logic to its own class
